flytt tester fra TilkjentYtelseUtilsTest.kt til TilkjentYtelseGeneratorTest
